+++ packages/theme_tailor/lib/src/model/tailor_annotation_data.dart
@@ -6,11 +6,13 @@ class TailorAnnotationData {
     required this.themes,
     required this.encoders,
     required this.requireStaticConst,
+    required this.generateStaticGetters,
     required this.themeGetter,
   });
 
   final List<String> themes;
   final Map<String, ThemeEncoderData> encoders;
   final bool requireStaticConst;
+  final bool generateStaticGetters;
   final ThemeGetter themeGetter;
 }