+++ packages/theme_tailor_annotation/lib/src/theme_tailor.dart
@@ -23,6 +23,7 @@ class Tailor {
     this.themeGetter,
     this.encoders,
     this.requireStaticConst,
+    this.generateStaticGetters,
   });
 
   factory Tailor.fromJson(Map json) => _$TailorFromJson(json);
@@ -86,4 +87,35 @@ class Tailor {
   /// }
   /// ```
   final bool? requireStaticConst;
+
+  /// When set to true, this option generates static getters that allow updating
+  /// theme properties on hot reload. The getters will conditionally return either
+  /// the theme itself (if kDebugMode == true) or the final theme otherwise.
+  /// Any property of the annotated class that should be updated on hot reload
+  /// must be either a getter or a const variable!
+  ///
+  /// To use this option, make sure to import `package:flutter/foundation.dart`
+  ///
+  /// Here's an example usage:
+  /// ```dart
+  /// const lightColor = Color(0xFFA1B2C3);
+  /// const darkColor = Color(0xFF123ABC);
+  ///
+  /// @tailor(generateStaticGetters: true)
+  /// class MyTheme {
+  /// // This is correct
+  /// static const color1 = [lightColor, darkColor];
+  /// static List<Color> get color2 => [lightColor, darkColor];
+  ///
+  /// // This is incorrect
+  ///  static List<Color> color3 = [lightColor, darkColor];
+  /// ```
+  /// In this example, the class defines several
+  /// properties, including color1, which is a const variable, and color2,
+  /// which is a getter. Both of these properties will be updated on hot reload
+  /// when their values change.
+  ///
+  /// However, color3 is a non-const variable, so it won't be updated on hot
+  /// reload even if generateStaticGetters is set to true.
+  final bool? generateStaticGetters;
 }