+++ packages/theme_tailor_annotation/lib/src/theme_tailor_component.dart
@@ -12,5 +12,7 @@ class TailorComponent extends Tailor {
   const TailorComponent({
     super.themes,
     super.encoders,
+    super.requireStaticConst,
+    super.generateStaticGetters,
   }) : super(themeGetter: ThemeGetter.none);
 }