+++ packages/theme_tailor/CHANGELOG.md
@@ -1,5 +1,6 @@
 # 1.2.0-dev
-- Build configuration. It is possible to configure `themeGetter` and `requireStaticConst` annotation properties within the build.yaml with: theme_getter and require_static_const (all keys and values are snake_case)
+- Generate static getters, allowing the theme to be updated on hot reload.
+- Build configuration. It is possible to configure `themeGetter`, `requireStaticConst` and `generateStaticGetters` annotation properties within the build.yaml with: theme_getter and require_static_const (all keys and values for enums are snake_case)
 
 # 1.1.2
 - Fix: DiagnosticableTreeMixin not being generated if package:flutter/foundation.dart was proceeded by other Flutter import